Part 3 lab
Currently the Index method returns a string with a message that is
hard-coded in the controller class. Change the Index method to return a
View object.
The Index method above uses a view template to generate an HTML response
to the browser. Controller methods (also known as action methods), such
as the Index method above, generally return an  ActionResult (or a class
derived from  ActionResult), not primitive types like string.